[geometry/dev] Move test_utilities to test
test_png and test_tiff do not need to be available elsewhere
and can remain in the test folder.
