--- geometry/render/dev/test_utilities/BUILD.bazel
@@ -1,49 +1,0 @@
-# -*- python -*-
-# This file contains rules for Bazel; see drake/doc/bazel.rst.
-
-load(
-    "@drake//tools/skylark:drake_cc.bzl",
-    "drake_cc_googletest",
-    "drake_cc_library",
-    "drake_cc_package_library",
-)
-load("//tools/lint:lint.bzl", "add_lint_tests")
-
-package(default_visibility = ["//visibility:public"])
-
-drake_cc_package_library(
-    name = "test_utilities",
-    testonly = 1,
-    visibility = ["//visibility:public"],
-    deps = [
-        ":test_png",
-        ":test_tiff",
-    ],
-)
-
-drake_cc_library(
-    name = "test_png",
-    testonly = 1,
-    srcs = ["test_png.cc"],
-    hdrs = ["test_png.h"],
-    deps = [
-        "//common:unused",
-        "//systems/sensors:image",
-        "@gtest//:without_main",
-        "@libpng",
-    ],
-)
-
-drake_cc_library(
-    name = "test_tiff",
-    testonly = 1,
-    srcs = ["test_tiff.cc"],
-    hdrs = ["test_tiff.h"],
-    deps = [
-        "//systems/sensors:image",
-        "@gtest//:without_main",
-        "@libtiff",
-    ],
-)
-
-add_lint_tests()
